+++ packaging/whykey.nix
@@ -2,7 +2,7 @@
 
 rustPlatform.buildRustPackage {
   pname = "whykey";
-  version = "1.2.2";
+  version = "1.3.0";
 
   src = lib.cleanSource ../.;
   cargoLock.lockFile = ../Cargo.lock;